Move exchange utilities into separate module
